% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{assign_dist}
\alias{assign_dist}
\title{assign appropriate numeric coding for specified distribution}
\usage{
assign_dist(prior)
}
\arguments{
\item{prior}{the named list returned by the specific prior function
this code is taken from the rstanarm package}
}
\value{
list with all possible parameter values appropriately populated
}
\description{
assign appropriate numeric coding for specified distribution
}
