% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_constant_vars}
\alias{check_constant_vars}
\title{Check if any variables in a model frame are constants
(the exception is that a constant variable of all 1's is allowed)}
\usage{
check_constant_vars(mf)
}
\arguments{
\item{mf}{A model frame or model matrix}
}
\value{
If no constant variables are found mf is returned, otherwise an error
is thrown.
}
\description{
Check if any variables in a model frame are constants
(the exception is that a constant variable of all 1's is allowed)
}
