% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_data.R
\name{check_dups.stap_data}
\alias{check_dups.stap_data}
\title{checks for duplicates in stap,sap,tap designation}
\usage{
\method{check_dups}{stap_data}(object)
}
\description{
checks for duplicates in stap,sap,tap designation
}
