% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{check_exposure_plot}
\alias{check_exposure_plot}
\title{Exposure Plots}
\usage{
check_exposure_plot(samples, samples_2 = NULL, K = function(x, y)
  exp(-x/y), grid = seq(from = 0, to = 10, by = 0.01), q = c(0.025,
  0.975), Distance = TRUE)
}
\arguments{
\item{samples}{samples from the distribution of interest}

\item{samples_2}{samples from the distribution of interest for a 2D parameter NULL by default}

\item{K}{function with 2 arguments the first for distance/time and the second (and/or third) for evaluating a 2D exposure function}

\item{grid}{grid values to evaluate K on}

\item{q}{quantiles at which to evaluate the sample}

\item{Distance}{boolean value used to label horizontal axis of plot}
}
\description{
Exposure Plots
}
\details{
Useful for deciding whether you have the right prior 
for a spatial temporal parameters
}
