% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-example_model.R
\name{example_model}
\alias{example_model}
\title{Example model}
\format{Calling \code{example("example_model")} will run the model in the 
  Examples section, below, and the resulting stapreg object will then be
  available in the global environment. The \code{chains} and \code{iter}
  arguments are specified to make this example small in size. In practice,
  we recommend that they be left unspecified in order to use the default
  values (4 and 2000 respectively) or increased if there are convergence
  problems. The \code{cores} argument is optional and on a multicore system,
  the user may well want to set that equal to the number of chains being
  executed.}
\description{
A model for use in \pkg{rstap} examples.
}
\examples{
 ## following lines make example run faster
distdata <- subset(homog_longitudinal_bef_data[,c("subj_ID","measure_ID","class","dist")],subj_ID<=25)
timedata <- subset(homog_longitudinal_bef_data[,c("subj_ID","measure_ID","class","time")],subj_ID<=25)
timedata$time <- as.numeric(timedata$time)
subjdata <- subset(homog_longitudinal_subject_data,subj_ID<=25)
example_model <- 
  stap_glmer(y_bern ~ centered_income +  sex + centered_age + stap(Coffee_Shop) + (1|subj_ID),
                  family = binomial(link='logit'),
                  subject_data = subjdata,
                  distance_data = distdata,
                  time_data = timedata,
                  subject_ID = 'subj_ID',
                  group_ID = 'measure_ID',
                  prior_intercept = normal(location = 25, scale = 4, autoscale = F),
                  prior = normal(location = 0, scale = 4, autoscale=F),
                  prior_stap = normal(location = 0, scale = 4),
                  prior_theta = list(Coffee_Shop = list(spatial = log_normal(location = 1,
                                                                             scale = 1),
                                                         temporal = log_normal(location = 1,
                                                                               scale = 1))),
                  max_distance = 3, max_time = 50,
                  # chains, cores, and iter set to make the example small and fast
                  chains = 1,  
                  iter = 5E2, cores = 1)
example_model
}
\seealso{
The Longituinal \href{https://biostatistics4socialimpact.github.io/rstap/articles/longitudinal-I.html}{Vignette} for \code{stap_glmer}.
}
