% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_block.R
\name{extract_stap_data}
\alias{extract_stap_data}
\title{Extract stap data from formula and create stap_data object}
\usage{
extract_stap_data(formula)
}
\arguments{
\item{formula}{that designates model expression including stap covariates}
}
\description{
Extract stap data from formula and create stap_data object
}
