% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benvo-methods.R
\name{extract_tmat}
\alias{extract_tmat}
\title{Extract time vector for compressed row storage format}
\usage{
extract_tmat(benvo, term, component)
}
\arguments{
\item{benvo}{built environment object from \pkg{rbenvo}}

\item{term}{string of term in benvo}

\item{component}{string of benvo term component (e.g. Distance, Time, or Distance-Time)}
}
\description{
Extract time vector for compressed row storage format
}
