% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{family.stapreg}
\alias{family.stapreg}
\title{family method for stapreg objects}
\usage{
\method{family}{stapreg}(object, ...)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{family}}.}
}
\description{
family method for stapreg objects
}
\keyword{internal}
