% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{formula.stapreg}
\alias{formula.stapreg}
\title{formula method for stapreg objects}
\usage{
\method{formula}{stapreg}(x, ...)
}
\arguments{
\item{x}{A stapreg object.}

\item{...}{Can contain \code{fixed.only} and \code{random.only} arguments 
that both default to \code{FALSE}.}
}
\description{
formula method for stapreg objects
}
\keyword{internal}
