% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_block.R
\name{get_stap_code}
\alias{get_stap_code}
\title{Get stap coding from formula}
\usage{
get_stap_code(all_names, stap_covs)
}
\arguments{
\item{all_names}{character vector from calling all.names(formula)}
}
\value{
vector of length equal to number of staps + saps + taps
with the appropriate coding for each appropriate predictor
}
\description{
Get stap coding from formula
}
