% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_block.R
\name{get_stapless_formula}
\alias{get_stapless_formula}
\title{get_stapless_formula}
\usage{
get_stapless_formula(f)
}
\arguments{
\item{f}{formula from stap_glm}
}
\value{
formula without ~ stap() components
}
\description{
Get formula for typical covariates
}
