% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik.stapreg}
\alias{log_lik.stapreg}
\alias{log_lik}
\title{Pointwise log-likelihood matrix}
\usage{
\method{log_lik}{stapreg}(object, newdata = NULL, offset = NULL, ...)
}
\arguments{
\item{object}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{newdata}{An optional data frame of new data (e.g. holdout data) to use
when evaluating the log-likelihood. See the description of \code{newdata}
for \code{\link{posterior_predict}}.}

\item{offset}{A vector of offsets. Only required if \code{newdata} is
specified and an \code{offset} was specified when fitting the model.}

\item{...}{Currently ignored.}
}
\value{
A \eqn{S} by \eqn{N} matrix, where \eqn{S} is the size of the posterior  
  sample and \eqn{N} is the number of data points.
}
\description{
For models fit using MCMC, the \code{log_lik} method returns the
\eqn{S} by \eqn{N} pointwise log-likelihood matrix, where \eqn{S} is the size
of the posterior sample and \eqn{N} is the number of data points.
}
