% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{model.frame.stapreg}
\alias{model.frame.stapreg}
\title{model.frame method for stapreg objects}
\usage{
\method{model.frame}{stapreg}(formula, fixed.only = FALSE, ...)
}
\arguments{
\item{formula, ...}{See \code{\link[stats]{model.frame}}.}

\item{fixed.only}{See \code{\link[lme4]{model.frame.merMod}}.}
}
\description{
model.frame method for stapreg objects
}
\keyword{internal}
