% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{model.matrix.stapreg}
\alias{model.matrix.stapreg}
\title{model.matrix method for stapreg objects}
\usage{
\method{model.matrix}{stapreg}(object, subject_data)
}
\arguments{
\item{object, ...}{See \code{\link[stats]{model.matrix}}.}
}
\description{
model.matrix method for stapreg objects
}
\keyword{internal}
