% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{nsap}
\alias{nsap}
\title{Retrieves number of temporal aggregated predictors}
\usage{
nsap(object)
}
\arguments{
\item{object}{A fitted model object}
}
\value{
number of spatial aggregated predictors
}
\description{
Retrieves number of temporal aggregated predictors
}
\seealso{
\code{\link{nstap.stapreg}}
}
\keyword{internal}
