% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{nstap}
\alias{nstap}
\title{Retrieves number of Spatial-temporal aggregated predictors}
\usage{
nstap(object)
}
\value{
number of spatial temporal aggregated predictors
}
\description{
Retrieves number of Spatial-temporal aggregated predictors
}
\seealso{
\code{\link{nstap.stapreg}}
}
\keyword{internal}
