% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{ntap}
\alias{ntap}
\title{Retrieves number of temporal aggregated predictors}
\usage{
ntap(object)
}
\value{
number of temporal aggregated predictors
}
\description{
Retrieves number of temporal aggregated predictors
}
\seealso{
\code{\link{nstap.stapreg}}
}
\keyword{internal}
