% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_stap}
\alias{plot_stap}
\title{plot_stap}
\usage{
plot_stap(object, pars, regex_pars)
}
\arguments{
\item{object}{a fitted stapreg object}

\item{pars}{a specific parameter to plot}

\item{regex_pars}{a string to run through regex  for matching}
}
\value{
a plot with the appropriate  weight function and scale
}
\description{
plot_stap
}
