% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap.R
\name{sap}
\alias{sap}
\title{Spatial  Aggregated Predictor formula function}
\usage{
sap(name = "", s_function = "wei")
}
\arguments{
\item{name}{string name}

\item{s_function}{one of "wei","exp","erf"}

\item{t_function}{one of "wei", "exp","erf"}
}
\description{
Spatial  Aggregated Predictor formula function
}
