% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{set_sampling_args}
\alias{set_sampling_args}
\title{Set arguments for sampling}
\usage{
set_sampling_args(object, prior, user_dots = list(),
  user_adapt_delta = NULL, user_max_tree_depth = 10, ...)
}
\arguments{
\item{object}{The stanfit object to use for sampling.}

\item{prior}{Prior distribution list (can be NULL).}

\item{user_dots}{The contents of \code{...} from the user's call to
the \code{stan_*} modeling function.}

\item{user_adapt_delta}{The value for \code{adapt_delta} specified by the
user.}

\item{...}{Other arguments to \code{\link[rstan]{sampling}} not coming from
\code{user_dots} (e.g. \code{data}, \code{pars}, \code{init}, etc.)}
}
\value{
A list of arguments to use for the \code{args} argument for
  \code{do.call(sampling, args)}.
}
\description{
Prepare a list of arguments to use with \code{rstan::sampling} via
\code{do.call}.
}
