% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap.R
\name{stap}
\alias{stap}
\title{Spatial Temporal Aggregated Predictor formula function}
\usage{
stap(name = "", s_function = "wei", t_function = "wei")
}
\arguments{
\item{name}{string name}

\item{s_function}{one of "wei","exp","erf"}

\item{t_function}{one of "wei", "exp","erf"}
}
\description{
Spatial Temporal Aggregated Predictor formula function
}
