% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_data.R
\name{stap_data}
\alias{stap_data}
\title{Create a stap_data object}
\usage{
stap_data(object)
}
\arguments{
\item{object}{a named list of objects containing information about the staps for a given model}
}
\value{
an object of class "stap_data"
}
\description{
Create a stap_data object
}
