% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_glm.R
\name{stap_glm}
\alias{stap_glm}
\title{Fitting Generalized Linear STAP models}
\usage{
stap_glm(formula, family = gaussian(), subject_data = NULL,
  distance_data = NULL, time_data = NULL, id_key = NULL, max_distance,
  weights, subset, na.action = NULL, offset = NULL, model = TRUE,
  z = FALSE, y = TRUE, x = FALSE, contrasts = NULL, ...,
  prior = normal(), prior_intercept = normal(), prior_stap = normal(),
  prior_theta = normal(location = 30, scale = 10),
  prior_aux = cauchy(location = 0L, scale = 5L), adapt_delta = NULL)
}
\arguments{
\item{family}{Same as \code{\link[stats]{glm}} for gaussian, binomial, and poisson}

\item{id_key}{name of column to join on between subject_data and bef_data}

\item{max_distance}{the inclusion distance; upper bound for all elements of dists_crs}
}
\description{
Fitting Generalized Linear STAP models
}
\details{
The \code{stap_glm} function is similar in syntax to
\code{\link[rstanarm]{stan_glm}} except instead of performing full bayesian
inference for a generalized linear model stap_glm incorporates spatial
as detailed in in --need to add citation --
}
