% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{stap_prior}
\alias{stap_prior}
\title{Stap Prior Function}
\usage{
stap_prior(
  intercept = normal(),
  delta = normal(),
  beta = normal(),
  theta = log_normal(),
  auxiliary = cauchy()
)
}
\arguments{
\item{intercept}{prior for intercept variable}

\item{delta}{prior for standard regression coefficients}

\item{beta}{prior for stap regression coefficients}

\item{theta}{prior for spatio-temporal scale regression coefficients}

\item{auxiliary}{prior for auxiliary variance}
}
\description{
Stap Prior Function
}
