% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg.R
\name{stapreg}
\alias{stapreg}
\title{Create a stapreg object}
\usage{
stapreg(object)
}
\arguments{
\item{object}{A list provided by one of the \code{stap_*} modeling functions.}
}
\value{
A stanreg object
}
\description{
Create a stapreg object
}
