% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapspec-methods.R
\name{stapspec-methods}
\alias{stapspec-methods}
\alias{get_Q}
\title{Stap Specification Methods}
\usage{
get_Q(x)
}
\arguments{
\item{x}{stapspec object}
}
\description{
Stap Specification Methods
}
\details{
methods for stapspec objects. Function names are typically self explanatory
}
