% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap.R
\name{tap}
\alias{tap}
\title{Temporal Aggregated Predictor formula function}
\usage{
tap(name = "", t_function = "wei")
}
\arguments{
\item{name}{string name}

\item{t_function}{one of "wei", "exp","erf"}

\item{s_function}{one of "wei","exp","erf"}
}
\description{
Temporal Aggregated Predictor formula function
}
