% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R
\name{terms.stapreg}
\alias{terms.stapreg}
\title{terms method for stapreg objects}
\usage{
\method{terms}{stapreg}(x, ..., fixed.only = TRUE, random.only = FALSE)
}
\arguments{
\item{x, fixed.only, random.only, ...}{See lme4:::terms.merMod.}
}
\description{
terms method for stapreg objects
}
\keyword{internal}
