% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_distancedata}
\alias{validate_distancedata}
\title{Validate distance_data}
\usage{
validate_distancedata(distance_data, max_distance)
}
\arguments{
\item{distance_data}{User's distance_data argument}
}
\value{
If no error is thrown, the column index
for the distance data is returned. If no distance_data is supplied NULL type returned.
}
\description{
Make sure that data is a data frame.
}
