% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_family}
\alias{validate_family}
\title{Check family argument}
\usage{
validate_family(f)
}
\value{
If no error is thrown than either \code{f} itself is returned
(if already a family) or the family object created from \code{f} is 
returned if \code{f} a string or function. Code adapted from \pkg{rstanarm}.
}
\description{
Check family argument
}
