% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_newdata}
\alias{validate_newdata}
\title{Validate newdata argument for posterior_predict, log_lik, etc.}
\usage{
validate_newdata(x)
}
\arguments{
\item{x}{User's 'newdata' argument}
}
\value{
Either NULL or a data frame
}
\description{
Doesn't check if the correct variables are included (that's done in pp_data),
just that newdata is either NULL or a data frame with no missing values. Also
drops any unused dimensions in variables (e.g. a one column matrix inside a
data frame is converted to a vector).
}
