% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_timedata}
\alias{validate_timedata}
\title{Validate time_data}
\usage{
validate_timedata(time_data)
}
\arguments{
\item{time_data}{User's time_data argument}
}
\value{
If no error is thrown, the index corresponding to the column
holding the time data is returned. If no time_data is supplied NULL type returned.
}
\description{
Make sure that time_data is a data frame, return time column index.
}
