% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_weights}
\alias{validate_weights}
\title{Check weights argument}
\usage{
validate_weights(w)
}
\arguments{
\item{w}{The \code{weights} argument specified by user or the result of
calling \code{model.weights} on a model frame.}
}
\value{
If no error is thrown then \code{w} is returned.
}
\description{
Check weights argument
}
