% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stapreg-methods.R, R/waic.R
\name{waic}
\alias{waic}
\alias{waic.stapreg}
\title{Method for calculating Widely Applicable Information Criteria}
\usage{
waic(x)

\method{waic}{stapreg}(x)
}
\arguments{
\item{x}{a stapreg object}
}
\description{
Method for calculating Widely Applicable Information Criteria

WAIC
}
