% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_block.R
\name{extract_stap_data}
\alias{extract_stap_data}
\title{extract stap data from formula and create stapcov object}
\usage{
extract_stap_data(formula)
}
\arguments{
\item{formula}{that designates model expression including stap covariates}
}
\value{
stap_list - a list for each stap containing the covariate name, stap type
       weight function and log indicator
}
\description{
extract stap data from formula and create stapcov object
}
