% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nlist}
\alias{nlist}
\title{create a named list using specified names or, if names are omitted using the
names of the objects in the list}
\usage{
nlist(...)
}
\description{
create a named list using specified names or, if names are omitted using the
names of the objects in the list
}
