% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstap-package.R
\docType{package}
\name{rstap-package}
\alias{rstap-package}
\alias{rstap}
\title{The 'rstap' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
