% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstap-package.R
\docType{package}
\name{rstap-package}
\alias{rstap-package}
\alias{rstap}
\title{The 'rstap' package.}
\description{
rstap is a package that implements spatial-temporal aggregated predictor functions in R. This allows for the modeling of features impact on measured subjects that can be related either through space or time.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
