% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_glm.fit.R
\name{stap_glm.fit}
\alias{stap_glm.fit}
\title{Fitting Generalized Linear STKAP models}
\usage{
stap_glm.fit(y, z, dists_crs, u, max_distance = 3L, weights = rep(1,
  NROW(y)), offset = rep(0, NROW(y)), family = stats::gaussian(), ...,
  prior = normal(), prior_intercept = normal(), prior_stap = normal(),
  group = list(), prior_theta = list(theta_one = normal()),
  prior_aux = cauchy(location = 0L, scale = 5L), adapt_delta = NULL)
}
\arguments{
\item{y}{n length vector or n x 2 matrix of outcomes}

\item{dists_crs}{q x M matrix of distances between outcome observations and
environmental features where q is the number of spatial covariates,
 and M is the maximum number of environmental features amongst all q features}

\item{max_distance}{the upper bound of distance for which all}

\item{Z}{n x p design matrix of subject specific covariates}
}
\description{
Fitting Generalized Linear STKAP models
}
