% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_glm.R, R/stap_glm.fit.R
\name{stap_glm}
\alias{stap_glm}
\alias{stap_lm}
\alias{stap_glm.fit}
\title{Bayesian generalized spatial-temporal aggregated predictor(STAP) models via Stan}
\usage{
stap_glm(formula, family = gaussian(), benvo, prior_list = stap_priors(), ...)

stap_lm(formula, benvo, prior = stap_prior(), ...)

stap_glm.fit(
  y,
  z,
  dists_crs,
  u_s,
  times_crs,
  u_t,
  stap_codes,
  family = stats::gaussian(),
  ...,
  prior_list = stap_prior(),
  optimize = FALSE,
  vb = FALSE
)
}
\arguments{
\item{formula}{Similar as for \code{\link[stats]{glm}} with the addition of \code{stap},\code{sap} \code{tap} terms as needed}

\item{family}{distributional family - only binomial gaussian or poisson currently allowed}

\item{benvo}{built environment \code{\link[rbenvo]{benvo}} object from the \code{rbenvo} package containing the relevant subject-BEF data}

\item{prior_list}{see \code{\link{stap_glm}} and \code{\link{prior_stap}} for more information}

\item{...}{optional arguments passed to the sampler - e.g. iter,warmup, etc.}

\item{y}{n length vector or n x 2 matrix of outcomes}

\item{z}{n x p design matrix of subject specific covariates}

\item{dists_crs}{(q_s+q_st) x M matrix of distances between outcome 
observations and built environment features with a hypothesized spatial scale}

\item{u_s}{n x (q *2) matrix of compressed row storage array indices for dists_crs}

\item{times_crs}{(q_t+q_st) x M matrix of times where the outcome observations
were exposed to the built environment features with a hypothesized temporal scale}

\item{u_t}{n x (q*2) matrix of compressed row storage array  indices for times_crs}

\item{offset, weights}{Same as \code{\link[stats]{glm}}.}

\item{contrasts}{Same as \code{\link[stats]{glm}}, but
rarely specified.}

\item{spec}{stap model specification - of class stapspec}

\item{weight_functions}{a Q x 2 matrix with integers coding the appropriate weight function for each STAP}

\item{group}{list of of group terms from \code{lme4::glmod}}
}
\value{
A \link[=stapreg-objects]{stapreg} object is returned 
for \code{stap_glm}.

A \link[=stanfit-class]{stapfit} object (or a slightly modified 
  stapfit object) is returned if \code{stan_glm.fit} is called directly.
}
\description{
Generalized linear modeling with spatial temporal aggregated predictors using
prior distributions for the coefficients, intercept, spatial-temporal scales, and auxiliary parameters.
}
\details{
The \code{stap_glm} function is similar in syntax to
\code{\link[rstanarm]{stan_glm}} except instead of performing full bayesian
inference for a generalized linear model stap_glm incorporates spatial-temporal covariates
}
\section{Functions}{
\itemize{
\item \code{stap_glm.fit}: 
}}

\examples{

fit_glm <- stap_glm(formula = y ~ sex + sap(Fast_Food),
				   benvo = FFR_example,
                     family = gaussian(link = 'identity'),
					prior = stap_priors()
                     chains = 1, iter = 300, # for speed of example only
                     refresh = -1, verbose = FALSE) 

}
\references{
Gelman, A. and Hill, J. (2007). \emph{Data Analysis Using
  Regression and Multilevel/Hierarchical Models.} Cambridge University Press,
  Cambridge, UK.

Muth, C., Oravecz, Z., and Gabry, J. (2018)
User-friendly Bayesian regression modeling: A tutorial with rstanarm and shinystan.
\emph{The Quantitative Methods for Psychology}. 14(2), 99--119.
\url{https://www.tqmp.org/RegularArticles/vol14-2/p099/p099.pdf}
}
\seealso{
\code{\link{stapreg-methods}} and 
\code{\link[stats]{glm}}.

The various vignettes for \code{stap_glm} at
  \url{https://biostatistics4socialimpact.github.io/rstap/articles} and the \href{http://arxiv.org/abs/1812.10208}{preprint} article.
}
