% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_lm.R
\name{stap_lm}
\alias{stap_lm}
\title{Fitting Linear STKAP models}
\usage{
stap_lm(y, Z, dists_csr, u, max_distance = 3, prior = normal(),
  prior_intercept = normal(location = 25, scale = 5), prior_stap = normal(),
  prior_theta = list(theta_one = normal(location = 1.5, scale = 0.5)),
  prior_sigma = cauchy(location = 0, scale = 5), ...)
}
\arguments{
\item{y}{vector of continuous outcomes}

\item{Z}{n*p design matrix of subject specific covariates}

\item{dists_csr}{distance crs array}

\item{u}{crs index array}

\item{max_distance}{inclusion distance}
}
\description{
Fitting Linear STKAP models
}
