% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap_lm.R
\name{stap_lm}
\alias{stap_lm}
\title{Fitting Linear STKAP models}
\usage{
stap_lm(y, Z, dists_csr, u, max_distance = 3,
  prior_intercept = normal(location = 25, scale = 5),
  prior_beta_one = normal(location = 0, scale = 5),
  prior_beta_two = normal(location = 0, scale = 5),
  prior_theta = list(theta_one = normal(location = 1.5, scale = 0.5)),
  prior_sigma = cauchy(location = 0, scale = 5), chains = 1, iter = 500,
  warmup = floor(iter/2), thin = 1, cores = getOption("mc.cores", 1L),
  seed = Sys.Date(), control = list(...), ...)
}
\arguments{
\item{y}{vector of continuous outcomes}

\item{Z}{n*p design matrix of subject specific covariates}

\item{X}{vector of environmental features integer coding}
}
\description{
Fitting Linear STKAP models
}
