% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.stapreg.R, R/log_lik.R,
%   R/misc.R, R/plots.R, R/posterior_interval.R, R/predictive_error.R,
%   R/print-and-summary.R, R/stapreg-objects.R
\name{stapreg-objects}
\alias{stapreg-objects}
\alias{as.matrix.stapreg}
\alias{log_lik.stapreg}
\alias{log_lik}
\alias{get_y}
\alias{plot.stapreg}
\alias{pairs.stapreg}
\alias{pairs_style_np}
\alias{pairs_condition}
\alias{posterior_interval.stapreg}
\alias{posterior_interval}
\alias{predictive_error.stapreg}
\alias{predictive_error}
\alias{print.stapreg}
\title{Extract the posterior sample}
\usage{
\method{as.matrix}{stapreg}(x, ..., pars = NULL, regex_pars = NULL)

\method{log_lik}{stapreg}(object, newsubjdata = NULL,
  newdistdata = NULL, newtimedata = NULL, offset = NULL, ...)

get_y(object, ...)

\method{plot}{stapreg}(x, plotfun = "intervals", pars = NULL,
  regex_pars = NULL, ...)

\method{pairs}{stapreg}(x, pars = NULL, regex_pars = NULL,
  condition = pairs_condition(nuts = "accept_stat__"), ...)

\method{posterior_interval}{stapreg}(object, prob = 0.9,
  type = "central", pars = NULL, regex_pars = NULL, ...)

\method{predictive_error}{stapreg}(object, newsubjdata = NULL,
  newdistdata = NULL, newtimedata = NULL, draws = NULL,
  re.form = NULL, seed = NULL, offset = NULL, ...)

\method{print}{stapreg}(x, digits = 1, ...)
}
\arguments{
\item{x}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{...}{Ignored.}

\item{pars}{An optional character vector of parameter names.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}. Currently, all
functions that accept a \code{regex_pars} argument ignore it for models fit
using optimization.}

\item{object}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{newsubjdata}{Optionally, a data frame of the subject-specific data
in which to look for variables with which to predict.
If omitted, the original datasets are used. If \code{newsubjdata}
is provided and any variables were transformed (e.g. rescaled) in the data
used to fit the model, then these variables must also be transformed in
\code{newsubjdata}.  Also see the Note
section below for a note about using the \code{newsubjdata} argument with with
binomial models.}

\item{newdistdata}{If newsubjdata is provided a data frame of the subject-distance
must also be given for models with a spatial component - can be the same as original distance_dataframe}

\item{newtimedata}{If newsubjdata is provided, a data frame of the subject-time data
must also be given for models with a temporal component}

\item{offset}{A vector of offsets. Only required if \code{newsubjdata} is
specified and an \code{offset} was specified when fitting the model.}

\item{plotfun}{A character string naming the \pkg{bayesplot} 
\link[bayesplot]{MCMC} function to use. The default is to call
\code{\link[bayesplot]{mcmc_intervals}}. \code{plotfun} can be specified
either as the full name of a \pkg{bayesplot} plotting function (e.g.
\code{"mcmc_hist"}) or can be abbreviated to the part of the name following
the \code{"mcmc_"} prefix (e.g. \code{"hist"}). To get the names of all
available MCMC functions see \code{\link[bayesplot]{available_mcmc}}.}

\item{condition}{Same as the \code{condition} argument to 
\code{\link[bayesplot]{mcmc_pairs}} except the \emph{default is different}
for \pkg{rstap} models. By default, the \code{mcmc_pairs} function in
the \pkg{bayesplot} package plots some of the Markov chains (half, in the
case of an even number of chains) in the panels above the diagonal and the
other half in the panels below the diagonal. However since we know that 
\pkg{rstap} models were fit using Stan (which \pkg{bayesplot} doesn't 
assume) we can make the default more useful by splitting the draws 
according to the \code{accept_stat__} diagnostic. The plots below the 
diagonal will contain realizations that are below the median 
\code{accept_stat__} and the plots above the diagonal will contain 
realizations that are above the median \code{accept_stat__}. To change this
behavior see the documentation of the \code{condition} argument at 
\code{\link[bayesplot]{mcmc_pairs}}.}

\item{prob}{A number \eqn{p \in (0,1)}{p (0 < p < 1)} indicating the desired
probability mass to include in the intervals. The default is to report
\eqn{90}\% intervals (\code{prob=0.9}) rather than the traditionally used
\eqn{95}\% (see Details).}

\item{type}{The type of interval to compute. Currently the only option is
\code{"central"} (see Details). A central \eqn{100p}\%
interval is defined by the \eqn{\alpha/2} and \eqn{1 - \alpha/2} quantiles,
where \eqn{\alpha = 1 - p}.}

\item{digits}{Number of digits to use for formatting numbers.}

\item{...}{Currently ignored.}

\item{object}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{...}{Other arguments passed to methods. For a \code{stanmvreg} object
this can be an integer \code{m} specifying the submodel.}

\item{x}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{pars}{An optional character vector of parameter names.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}. Currently, all
functions that accept a \code{regex_pars} argument ignore it for models fit
using optimization.}

\item{...}{Additional arguments to pass to \code{plotfun} for customizing the
plot. These are described on the help pages for the individual plotting 
functions. For example, the arguments accepted for the default
\code{plotfun="intervals"} can be found at
\code{\link[bayesplot]{mcmc_intervals}}.}

\item{x}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}. Currently, all
functions that accept a \code{regex_pars} argument ignore it for models fit
using optimization.}

\item{pars}{An optional character vetor of parameter names. All parameters 
are included by default, but for models with more than just a few 
parameters it may be far too many to visualize on a small computer screen 
and also may require substantial computing time.}

\item{...}{Optional arguments passed to \code{\link[bayesplot]{mcmc_pairs}}. 
The \code{np}, \code{lp}, and \code{max_treedepth} arguments to 
\code{mcmc_pairs} are handled automatically by \pkg{rstap} and do not 
need to be specified by the user in \code{...}. The arguments that can be 
specified in \code{...} include \code{transformations}, \code{diag_fun},
\code{off_diag_fun}, \code{diag_args}, \code{off_diag_args},
and \code{np_style}. These arguments are
documented thoroughly on the help page for
\code{\link[bayesplot]{mcmc_pairs}}.}

\item{object}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{...}{Currently ignored.}

\item{pars}{An optional character vector of parameter names.}

\item{regex_pars}{An optional character vector of \link[=grep]{regular 
expressions} to use for parameter selection. \code{regex_pars} can be used
in place of \code{pars} or in addition to \code{pars}. Currently, all
functions that accept a \code{regex_pars} argument ignore it for models fit
using optimization.}

\item{object}{Either a fitted model object returned by one of the 
\pkg{rstap} modeling functions (a \link[=stapreg-objects]{stapreg 
object}) or, for the \code{"ppd"} method, a matrix of draws from the 
posterior predictive distribution returned by 
\code{\link{posterior_predict}}.}

\item{newsubjdata, newdistdata, newtimedata, draws, seed, offset, re.form}{Optional arguments passed to 
\code{\link{posterior_predict}}. For binomial models, please see the
\strong{Note} section below if \code{newsubjdata} will be specified.}

\item{...}{Currently ignored.}

\item{x}{A fitted model object returned by one of the 
\pkg{rstap} modeling functions. See \code{\link{stapreg-objects}}.}

\item{...}{Ignored.}
}
\value{
A matrix, data.frame, or array, the dimensions of which depend on
  \code{pars} and \code{regex_pars}, as well as the model and estimation
  algorithm (see the Description section above).

A \eqn{S} by \eqn{N} matrix, where \eqn{S} is the size of the posterior  
  sample and \eqn{N} is the number of data points.

For \code{get_x} and \code{get_z}, a matrix. For \code{get_y}, either
  a vector or a matrix, depending on how the response variable was specified.

Either a ggplot object that can be further customized using the
  \pkg{ggplot2} package, or an object created from multiple ggplot objects
  (e.g. a gtable object created by \code{\link[gridExtra]{arrangeGrob}}).

A matrix with two columns and as many rows as model parameters (or
  the subset of parameters specified by \code{pars} and/or
  \code{regex_pars}). For a given value of \code{prob}, \eqn{p}, the columns
  correspond to the lower and upper \eqn{100p}\% interval limits and have the
  names \eqn{100\alpha/2}\% and \eqn{100(1 - \alpha/2)}\%, where \eqn{\alpha
  = 1-p}. For example, if \code{prob=0.9} is specified (a \eqn{90}\%
  interval), then the column names will be \code{"5\%"} and \code{"95\%"},
  respectively.

A \code{draws} by \code{nrow(newdata)} matrix. If \code{newdata} is 
  not specified then it will be \code{draws} by \code{nobs(object)}.

Returns \code{x}, invisibly.
}
\description{
For models fit using MCMC  the posterior sample ---the post-warmup 
draws from the posterior distribution--- can be extracted 
from a fitted model object as a matrix, data frame, or array. The 
\code{as.matrix} and \code{as.data.frame} methods merge all chains together, 
whereas the \code{as.array} method keeps the chains separate.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

For models fit using MCMC, the \code{log_lik} method returns the
\eqn{S} by \eqn{N} pointwise log-likelihood matrix, where \eqn{S} is the size
of the posterior sample and \eqn{N} is the number of data points.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

The \code{plot} method for \link{stapreg-objects} provides a convenient 
interface to the \link[bayesplot]{MCMC} module in the \pkg{\link{bayesplot}} 
package for plotting MCMC draws and diagnostics. It is also straightforward 
to use the functions from the \pkg{bayesplot} package directly rather than 
via the \code{plot} method. Examples of both methods of plotting are given
below.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

Interface to \pkg{bayesplot}'s \code{\link[bayesplot]{mcmc_pairs}} function 
for use with \pkg{rstap} models. Be careful not to specify too
many parameters to include or the plot will be both hard to read and slow to
render.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

For models fit using MCMC,  the \code{posterior_interval} 
function computes Bayesian posterior uncertainty
intervals. These intervals are often referred to as \emph{credible}
intervals, but we use the term \emph{uncertainty} intervals to highlight the
fact that wider intervals correspond to greater uncertainty.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

This is a convenience function for computing \eqn{y - y^{rep}}{y - yrep} 
(in-sample, for observed \eqn{y}) or \eqn{y - \tilde{y}}{y - ytilde} 
(out-of-sample, for new or held-out \eqn{y}). The method for stapreg objects 
calls \code{\link{posterior_predict}} internally, whereas the method for
objects with class \code{"ppd"} accepts the matrix returned by 
\code{posterior_predict} as input and can be used to avoid multiple calls to 
\code{posterior_predict}.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

The \code{print} method for stapreg objects displays a compact summary of the
fitted model. See the \strong{Details} section below for descriptions of the
different components of the printed output. For additional summary statistics
and diagnostics use the \code{\link[=summary.stapreg]{summary}} method.

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
and several additional components depending on the model and estimation 
algorithm. \cr
\cr

The \pkg{rstap} model-fitting functions return an object of class 
\code{'stapreg'}, which is a list containing at a minimum the components listed 
below. Each \code{stapreg} object will also have additional classes (e.g. 'glm')
\cr
}
\details{
\subsection{Interpretation}{
Unlike for a frenquentist confidence interval, it is valid to say that,
conditional on the data and model, we believe that with probability \eqn{p}
the value of a parameter is in its \eqn{100p}\% posterior interval. This
intuitive interpretation of Bayesian intervals is often erroneously applied
to frequentist confidence intervals. See Morey et al. (2015) for more details
on this issue and the advantages of using Bayesian posterior uncertainty
intervals (also known as credible intervals).
}
\subsection{Default 90\% intervals}{
We default to reporting \eqn{90}\% intervals rather than \eqn{95}\% intervals
for several reasons:
\itemize{
 \item Computational stability: \eqn{90}\% intervals are more stable than
 \eqn{95}\% intervals (for which each end relies on only \eqn{2.5}\% of the
 posterior draws). \item Relation to Type-S errors (Gelman and Carlin, 2014):
 \eqn{95}\% of the mass in a \eqn{90}\% central interval is above the lower
 value (and \eqn{95}\% is below the upper value). For a parameter
 \eqn{\theta}, it is therefore easy to see if the posterior probability that
 \eqn{\theta > 0} (or \eqn{\theta < 0}) is larger or smaller than \eqn{95}\%.
}
Of course, if \eqn{95}\% intervals are desired they can be computed by
specifying \code{prob=0.95}.
}
\subsection{Types of intervals}{
Currently \code{posterior_interval} only computes central intervals because
other types of intervals are rarely useful for the models that \pkg{rstap}
can estimate. Additional possibilities may be provided in future releases as
more models become available.
}

\subsection{Point estimates}{
Point estimates are medians computed from simulations.
For models fit using MCMC (\code{"sampling"}) the posterior
sample is used.  The point estimates reported are the same as the values
returned by \code{\link[=coef.stapreg]{coef}}.
}
\subsection{Uncertainty estimates (MAD_SD)}{
The standard deviations reported (labeled \code{MAD_SD} in the print output)
are computed from the same set of draws described above and are proportional
to the median absolute deviation (\code{\link[stats]{mad}}) from the median.
Compared to the raw posterior standard deviation, the MAD_SD will be
more robust for long-tailed distributions. These are the same as the values
returned by \code{\link[=se.stapreg]{se}}.
}
\subsection{Additional output}{
\itemize{
\item The median and MAD_SD are also reported for \code{mean_PPD}, the sample
average posterior predictive distribution of the outcome. This is useful as a
quick diagnostic. A useful heuristic is to check if \code{mean_PPD} is
plausible when compared to \code{mean(y)}. If it is plausible then this does
\emph{not} mean that the model is good in general (only that it can reproduce
the sample mean), however if \code{mean_PPD} is implausible then it is a sign
that something is wrong (severe model misspecification, problems with the
data, computational issues, etc.).

\item For GLMs with group-specific terms (see \code{\link{stap_glmer}}) the printed 
output also shows point estimates of the standard deviations of the group 
effects (and correlations if there are both intercept and slopes that vary by
group).

}
}
}
\note{
The \strong{Note} section in \code{\link{posterior_predict}} about 
  \code{newdata} for binomial models also applies for
  \code{predictive_error}, with one important difference. For
  \code{posterior_predict} if the left-hand side of the model formula is 
  \code{cbind(successes, failures)} then the particular values of 
  \code{successes} and \code{failures} in \code{newdata} don't matter, only 
  that they add to the desired number of trials. \strong{This is not the case
  for} \code{predictive_error}. For \code{predictive_error} the particular
  value of \code{successes} matters because it is used as \eqn{y} when
  computing the error.
}
\section{Elements for \code{stapreg} objects}{
   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution, the variational approximation, or the asymptotic 
  sampling distribution, depending on the estimation algorithm.
  }
  \item{\code{model,x,y}}{
  If requested, the the model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stapfit-class}} returned by RStan and a
  matrix of various summary statistics from the stapfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}

   
\describe{
  \item{\code{coefficients}}{
  Point estimates, as described in \code{\link{print.stapreg}}.
  }
  \item{\code{ses}}{
  Standard errors based on \code{\link[stats]{mad}}, as described in
  \code{\link{print.stapreg}}.
  }
  \item{\code{residuals}}{
  Residuals of type \code{'response'}.
  }
  \item{\code{fitted.values}}{
  Fitted mean values. For GLMs the linear predictors are transformed by the
  inverse link function.
  }
  \item{\code{linear.predictors}}{
  Linear fit on the link scale. For linear models this is the same as
  \code{fitted.values}.
  }
  \item{\code{covmat}}{
  Variance-covariance matrix for the coefficients based on draws from the
  posterior distribution. 
  }
  \item{\code{model,x,y,z}}{
  If requested, the latent samples,  model frame, model matrix and response variable used, 
  respectively.
  }
  \item{\code{family}}{
  The \code{\link[stats]{family}} object used.
  }
  \item{\code{call}}{
  The matched call.
  }
  \item{\code{formula}}{
  The model \code{\link[stats]{formula}}.
  }
  \item{\code{data,offset,weights}}{
  The \code{data}, \code{offset}, and \code{weights} arguments.
  }
  \item{\code{prior.info}}{
  A list with information about the prior distributions used.
  }
  \item{\code{stapfit,stan_summary}}{
  The object of \code{\link[rstan]{stanfit-class}} returned by RStan and a
  matrix of various summary statistics from the stanfit object.
  }
  \item{\code{rstan_version}}{
  The version of the \pkg{rstan} package that was used to fit the model.
  }
}
}

\references{
Gabry, J., Simpson, D., Vehtari, A., Betancourt, M., and Gelman,
  A. (2018). Visualization in Bayesian workflow. \emph{Journal of the Royal
  Statistical Society Series A}, accepted for publication. arXiv preprint:
  \url{http://arxiv.org/abs/1709.01449}.

Gelman, A. and Carlin, J. (2014). Beyond power calculations:
assessing Type S (sign) and Type M (magnitude) errors. \emph{Perspectives on
Psychological Science}. 9(6), 641--51.

Morey, R. D., Hoekstra, R., Rouder, J., Lee, M. D., and Wagenmakers, E. 
(2016). The fallacy of placing confidence in confidence intervals. 
\emph{Psychonomic Bulletin & Review}. 23(1), 103--123.
}
\seealso{
\code{\link{stapreg-methods}}

\itemize{ 
  \item The vignettes in the \pkg{bayesplot} package for many examples.
  \item \code{\link[bayesplot]{MCMC-overview}} (\pkg{bayesplot}) for links to
  the documentation for all the available plotting functions.
  \item \code{\link[bayesplot]{color_scheme_set}} (\pkg{bayesplot}) to change
  the color scheme used for plotting.
  \item \code{\link{pp_check}} for graphical posterior predictive checks.
}

\code{\link{predictive_interval}} for predictive intervals.

\code{\link[=posterior_predict.stapreg]{posterior_predict}} to draw
  from the posterior predictive distribution without computing predictive
  errors.

\code{\link{summary.stapreg}}, \code{\link{stapreg-methods}}

\code{\link{stapreg-methods}}
}
\keyword{internal}
